+ fixed displayinfo scale reading from DBC
*Tnx to Kaldorei for:
+ Quest system fixed
+ Mail system fixed
+ Auction house fixed
+ Guild creation fixed
+ Tauren size fixed
+ Rename bug at FLAG_RESTSTATE fixed

+++ src/shared/Database/DBCStructure.h
@@ -138,7 +138,7 @@ struct CreatureDisplayInfoEntry
 {
     uint32      Displayid;                                  // 0
                                                             // 1-8,unused
-    float       scale;                                      // 9
+    int32       scale;                                      // 9
                                                             // 10-11,unused
 };
